Use Initialize
We can now move to the left. We're using the initialize function to
register for events (not the constructor) yipee!.

I just learned context is actually the application object. This might
make things a LOT easier for testing, so I'm gonna try that first before
I expand the tests to move right/up/down.